+++ AllReadyApp/Web-App/AllReady.UnitTest/Areas/Admin/Controllers/TaskControllerTests.cs
@@ -6,6 +6,7 @@ using AllReady.Areas.Admin.Controllers;
 using AllReady.Areas.Admin.Features.Tasks;
 using AllReady.Areas.Admin.ViewModels.Task;
 using AllReady.Areas.Admin.ViewModels.Validators;
+using AllReady.Areas.Admin.ViewModels.Validators.Task;
 using AllReady.Extensions;
 using AllReady.UnitTest.Extensions;
 using MediatR;
